requirements.txt
django>=3.0.8,<4.0
mysqlclient
# uwsgi

# SASS dynamic compiler
libsass
django_sass_processor
django-cors-headers
django-filter

# CRUD
git+https://github.com/silentsokolov/django-admin-rangefilter.git
design-django-theme
django-unical-bootstrap-italia
git+https://github.com/UniversitaDellaCalabria/django-organizational-area.git
pillow

# Rest
#djangorestframework
git+https://github.com/peppelinux/django-rest-framework@agid_oas3

markdown
# openAPI schema v3
uritemplate
pyyaml
cryptography

python-magic

django-ckeditor
